+++ src/main/java/com/anthonyponte/wallet/repository/ITipoRepository.java
@@ -1,4 +1,3 @@
-/** template created by jowsnunez --> https://www.github.com/JowsNunez * */
 package com.anthonyponte.wallet.repository;
 
 import com.anthonyponte.wallet.entity.Tipo;
@@ -6,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITipoRepository extends JpaRepository<Tipo, Integer> {
+public interface ITipoRepository extends JpaRepository<Tipo, Long> {
 }
