+++ src/main/java/com/anthonyponte/wallet/repository/ITransaccionRepository.java
@@ -1,4 +1,3 @@
-/** template created by jowsnunez --> https://www.github.com/JowsNunez * */
 package com.anthonyponte.wallet.repository;
 
 import com.anthonyponte.wallet.entity.Transaccion;
@@ -7,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITransaccionRepository extends JpaRepository<Transaccion, Integer> {
+public interface ITransaccionRepository extends JpaRepository<Transaccion, Long> {
   @Query("FROM Transaccion t WHERE t.cuenta.idCuenta = ?1 ORDER BY t.fecha")
-  public Iterable<Transaccion> findAll(Integer id);
+  public Iterable<Transaccion> findAll(Long id);
 }
