+++ src/main/java/com/anthonyponte/wallet/repository/IUsuarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.anthonyponte.wallet.entity.Usuario;
 
 @Repository
-public interface IUsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface IUsuarioRepository extends JpaRepository<Usuario, Long> {
     public Usuario findByUsername(String username);
 }
